moves the marker to the geocoder where it belongs
