+++ example/src/pages/Geocoder.tsx
@@ -1,44 +1,20 @@
-import React, { useState, useMemo } from 'react'
+import React from 'react'
 import 'leaflet/dist/leaflet.css'
-import { Map, TileLayer, Marker } from '@datapunt/react-maps'
+import { Map, TileLayer } from '@datapunt/react-maps'
 import { ViewerContainer } from '@datapunt/asc-ui'
 import Controls from '../Zoom'
 import { utils } from '../../../src'
-import DefaultMarkerIcon from '../DefaultMarkerIcon'
 import NonTiledLayer from '../NonTiledLayer'
-import Geocoder, {
-  pointQuery,
+import Geocoder, { getSuggestions, getAddressById } from '../Geocoder'
+
+const geocoderProps = {
   getSuggestions,
   getAddressById,
-} from '../Geocoder'
-
-const markerPosition = {
-  lat: 52.3731081,
-  lng: 4.8932945,
 }
 
 const GeocoderPage = () => {
-  const [marker, setMarker] = useState()
-  const [clickPointInfo, setClickPointInfo] = useState()
-
-  const geocoderProps = useMemo(
-    () => ({
-      marker,
-      clickPointInfo,
-      getSuggestions,
-      getAddressById,
-    }),
-    [clickPointInfo, marker],
-  )
-
   return (
     <Map
-      events={{
-        click: async e => {
-          const pointInfo = await pointQuery(e)
-          setClickPointInfo(pointInfo)
-        },
-      }}
       options={{
         center: [52.3731081, 4.8932945],
         zoom: 10,
@@ -59,21 +35,12 @@ const GeocoderPage = () => {
         topLeft={<Geocoder {...geocoderProps} />}
         bottomRight={<Controls />}
       />
-      <Marker
-        setInstance={setMarker}
-        args={[markerPosition]}
-        options={{
-          icon: DefaultMarkerIcon,
-          opacity: 0,
-        }}
-      />
       <TileLayer
         args={['https://{s}.data.amsterdam.nl/topo_rd/{z}/{x}/{y}.png']}
         options={{
           subdomains: ['acc.t1', 'acc.t2', 'acc.t3', 'acc.t4'],
           tms: true,
-          attribution:
-            '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors',
+          attribution: 'Kaartgegevens CC-BY-4.0 Gemeente Amsterdam',
         }}
       />
       <NonTiledLayer
